modified acc to review comments
